LayoutTests/http/tests/security/PerformanceEntryList-parent-only-first-navigation.html

<script>
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.dumpAsText();
    }
    window.addEventListener("message", (e)=>{
        var entries = performance.getEntriesByType("resource");
        alert(e.data + " " + entries.length + " " + entries[0].name);
        if (window.testRunner)
            testRunner.notifyDone();
    }, false);
</script>
<iframe src="resources/navigate-and-postMessage.html"></iframe>
